Move new tests to CallSiteTests and make them more specific to ServiceCallSite
